+++ integration_tests/upgrade-test.nix
@@ -28,6 +28,8 @@ let
     (fetchFlake "crypto-org-chain/chain-main" "b3226f06fd2a236f9957304c4d83b0ea06ed2604").default;
   released5_0 =
     (fetchFlake "crypto-org-chain/chain-main" "246ba80c2f0c7e11a0a7b483a349d177ffeb0a9d").default;
+  released6_0 =
+    (fetchFlake "crypto-org-chain/chain-main" "bcd5bcb30a4bac7a5939a712ac079dd631abf41b").default;
   current = pkgs.callPackage ../. { };
 in
 pkgs.linkFarm "upgrade-test-package" [
@@ -53,6 +55,10 @@ pkgs.linkFarm "upgrade-test-package" [
   }
   {
     name = "v6.0.0";
+    path = released6_0;
+  }
+  {
+    name = "v7.0.0";
     path = current;
   }
 ]